third_party/WebKit/LayoutTests/imported/wpt/shadow-dom/Document-prototype-currentScript-expected.txt

CONSOLE ERROR: line 2443: Uncaught Error: assert_equals: expected Element node <script id="outerScriptElement">

var outerScriptElement ... but got null
CONSOLE ERROR: line 2443: Uncaught Error: assert_equals: expected Element node <script id="outerScriptElement">

var outerScriptElement ... but got null
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_equals: expected Element node <script id="outerScriptElement">

var outerScriptElement ... but got null
PASS document.currentScript must not to be set to a script element in a shadow tree in open mode 
PASS document.currentScript must not to be set to a script element in a shadow tree in closed mode 
PASS document.currentScript must be set to a script element that loads an external script in a document tree 
PASS document.currentScript must be set to a script element that loads an external script in a document tree 
PASS document.currentScript must not be set to a script element that loads an external script in an open shadow tree 
PASS document.currentScript must not be set to a script element that loads an external script in a closed shadow tree 
PASS document.currentScript must be set to a script element that loads an external script that was in an open shadow tree and then removed 
PASS document.currentScript must be set to a script element that loads an external script that was in a closed shadow tree and then removed 
Harness: the test ran to completion.
